pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "plan-manager"
version = "0.5.0"
description = "AI-assisted planning (project management) with MCP interface"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "sgrade"}
]
keywords = ["planning", "project-management", "ai", "mcp", "model-context-protocol"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
requires-python = ">=3.11"
dependencies = [
    "PyYAML>=6.0",        # YAML file parsing for plan data
    "mcp[cli]",           # MCP SDK for server functionality
    "starlette",          # ASGI framework used by FastMCP
    "uvicorn[standard]",  # ASGI server for running the MCP server
    "pydantic>=2.0",      # Schema validation for story and Plan models
]

[project.urls]
Homepage = "https://github.com/sgrade/plan-manager"
Repository = "https://github.com/sgrade/plan-manager"
Issues = "https://github.com/sgrade/plan-manager/issues"

[project.scripts]
pm = "plan_manager.__main__:main"

[tool.uv]
link-mode = "copy"

[tool.hatch.build.targets.wheel]
packages = ["src/plan_manager"]
